Set up the storyboard
Need to start working on the data model. Once the data model
is in place, use the FavoriteActors app search method to return
the picked actor ID to the Actor Map view controller in a label.
This will also require the movie database API to be written as
well. This seems like the bulk of the backround stuff. Having this
in place will make the rest of the app go smoothly.