Clicking an actor now sends the actors info to the Map View
The ActorMapViewContoller was not sending itself as the delegate to
the ActorSearchVeiwController so the delegate was not sending the
picked actor back. This has been fixed. Picking an actor now sends
the id of the actor to the AMVC and the delgate searches the id and
initializes a dictionary with all of the actors attributes.

Need to use the dictionary to initialize an actor and have it
geolocate and place a pin at the hometown and append the actor to the actors array at the same time. If the actor shares a hometown, add some offset to the pin.